+++ buildscripts/bundle_full.sh
@@ -1,9 +1,9 @@
 # --------------------------------------------------
 
-if [ ! -f deps ]; then
+if [ ! -f "deps" ]; then
   sudo rm -r deps
 fi
-if [ ! -f prefix ]; then
+if [ ! -f "prefix" ]; then
   sudo rm -r prefix
 fi
 
@@ -12,8 +12,10 @@ fi
 
 # --------------------------------------------------
 
-rm scripts/ffmpeg.sh
-cp flavors/full.sh scripts/ffmpeg.sh
+if [ ! -f "scripts/ffmpeg" ]; then
+  rm scripts/ffmpeg.sh
+fi
+cp flavors/default.sh scripts/ffmpeg.sh
 
 # --------------------------------------------------
